userdashboard component is refractored to quizes
