+++ Project/leader-boards/src/components/User.js
@@ -1,12 +1,12 @@
 import React from "react";
 
-import UserDashboard from './UserDashboard'
+import Quizes from "./Quizes";
 
 function User(props) {
 	return (
 		<div>
 			Welcome user you can see all the quizes alloted to you in here
-			<UserDashboard quizes={props.quizes} />
+			<Quizes quizes={props.quizes} />
 		</div>
 	);
 }
