+++ Project/leader-boards/src/components/quizes.js
@@ -1,37 +1,42 @@
-export const quizes = [
-	{
-		quizid: 1608139744278,
-		title: "Some Floppy title",
-		score: -1,
-		attq: -1,
-		response: {},
-	},
-	{
-		quizid: 1608139746288,
-		title: "Some Floppy title",
-		score: -1,
-		attq: -1,
-		response: {},
-	},
-	{
-		quizid: 1608139748285,
-		title: "title",
-		score: -1,
-		attq: -1,
-		response: {},
-	},
-	{
-		quizid: 1608139750285,
-		title: "Title it seems",
-		score: -1,
-		attq: -1,
-		response: {},
-	},
-	{
-		quizid: 1608139752278,
-		title: "Go away title",
-		score: -1,
-		attq: -1,
-		response: {},
-	}
-];
+import React from "react";
+// import { quizes } from "./quizes";
+import { Link, Route, Switch, useRouteMatch } from "react-router-dom";
+import Quiz from "./Quiz";
+import { qtns } from "./questions";
+
+function QuizCard(props) {
+	return (
+		<section className='quiz-card'>
+			<h2>{props.quiz.title}</h2>
+			<Link
+				className='card-button'
+				to={`${props.curURL}/${props.quiz.quizid}`}
+				onClick={() => console.log("quiz card link clicked", Date())}
+			>
+				Take test
+			</Link>
+		</section>
+	);
+}
+function Quizes(props) {
+	const { path, url } = useRouteMatch();
+	const { quizes } = props;
+	return (
+		<>
+			<Switch>
+				<Route path={`${path}/:quizid`}>
+					<Quiz qtns={qtns} />
+				</Route>
+				<Route path={path}>
+					<div className='user-dashboard'>
+						{quizes.map((quz, id) => {
+							return <QuizCard key={id} quiz={quz} curURL={url} />;
+						})}
+					</div>
+				</Route>
+			</Switch>
+		</>
+	);
+}
+
+export default Quizes;
